Introduced WorkflowTestUtil to enable reuse of registration and clean up in all specifications
